cloud_run_data_vendor/requirements.txt
alembic==1.4.3
amqp==5.0.2
attrs==20.3.0
billiard==3.6.3.0
bump2version==1.0.1
bumpversion==0.6.0
CacheControl==0.12.6
cachetools==4.1.1
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.2.1
decorator==4.4.2
dnspython==1.16.0
eventlet==0.29.1
Faker==4.17.1
firebase-admin==4.4.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
flower==0.9.5
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.4.3
google-cloud-firestore==2.0.0
google-cloud-language==2.0.0
google-cloud-logging==1.15.1
google-cloud-secret-manager==1.0.0
google-cloud-storage==1.32.0
google-cloud-videointelligence==1.16.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
googlemaps==4.4.2
greenlet==0.4.17
grpc-google-iam-v1==0.12.3
grpcio==1.33.2
httplib2==0.18.1
humanize==3.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==5.0.2
libcst==0.3.13
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
moviepy==1.0.3
msgpack==1.0.0
mypy-extensions==0.4.3
numpy==1.19.4
nylas==4.12.1
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.4
pg8000==1.12.5
Pillow==8.0.1
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.13.0
psycopg2-binary==2.8.6
pyactiveresource==2.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.4.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-http-client==3.3.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sendgrid==6.4.7
ShopifyAPI==8.0.4
six==1.15.0
SQLAlchemy==1.3.20
starkbank-ecdsa==1.1.0
text-unidecode==1.3
tornado==6.1
tqdm==4.54.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.25.11
URLObject==2.4.3
validator-collection==1.5.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
